+++ src/main/kotlin/com/github/jensim/megamanipulator/actions/search/sourcegraph/SourcegraphSearchClient.kt
@@ -2,10 +2,10 @@ package com.github.jensim.megamanipulator.actions.search.sourcegraph
 
 import com.github.jensim.megamanipulator.actions.NotificationsOperator
 import com.github.jensim.megamanipulator.actions.search.SearchResult
-import com.github.jensim.megamanipulator.actions.search.SearchTypes
 import com.github.jensim.megamanipulator.http.HttpClientProvider
 import com.github.jensim.megamanipulator.settings.SearchHostSettings.SourceGraphSettings
-import com.intellij.notification.NotificationType
+import com.intellij.notification.NotificationType.ERROR
+import com.intellij.notification.NotificationType.WARNING
 import com.jetbrains.rd.util.printlnError
 import io.ktor.client.features.timeout
 import io.ktor.client.request.post
@@ -22,22 +22,27 @@ class SourcegraphSearchClient(
 
     suspend fun search(searchHostName: String, settings: SourceGraphSettings, search: String): Set<SearchResult> {
         try {
-            /* TODO
-             * Paginate on result
-             */
             val raw = rawSearch(searchHostName, settings, search)
-            val response: SearchTypes.GraphQLResponse = json.decodeFromString(raw.readText())
+            if ("${raw.status.value}".take(1) != "2") {
+                throw RuntimeException("${raw.status} ${raw.readText()}")
+            }
+            val response: SourcegraphTypes.GraphQLResponse = json.decodeFromString(raw.readText())
             response.errors?.let {
                 printlnError("ERROR $it")
+                notificationsOperator.show(title = "Error from SourceGraph search result", body = "More info in application logs", type = ERROR)
+            }
+            response.data?.search?.results?.alert?.let {
+                printlnError("ALERT $it")
+                notificationsOperator.show(title = "Alert from SourceGraph search result", body = "More info in application logs", type = WARNING)
             }
             return response.data?.search?.results?.results.orEmpty().mapNotNull {
-                internalNameToSearchResult(searchHostName, it.repository?.name)
+                internalNameToSearchResult(searchHostName, it.getRepoName())
             }.toSet()
         } catch (e: Exception) {
             notificationsOperator.show(
-                title = "Search failed",
+                title = "Search failed (More data in app logs)",
                 body = e.message ?: "Exception running search",
-                type = NotificationType.ERROR,
+                type = ERROR,
             )
             e.printStackTrace()
             return emptySet()
@@ -47,7 +52,7 @@ class SourcegraphSearchClient(
     private suspend fun rawSearch(searchHostName: String, settings: SourceGraphSettings, search: String): HttpResponse {
         val client = httpClientProvider.getClient(searchHostName, settings)
         return client.post("${settings.baseUrl}/.api/graphql?Search=") {
-            body = SearchTypes.GraphQlRequest(SearchTypes.SearchVaraibles(search))
+            body = SourcegraphTypes.GraphQlRequest(SourcegraphTypes.SearchVaraibles(search))
             timeout {
                 socketTimeoutMillis = 5 * 60 * 1000
                 requestTimeoutMillis = 5 * 60 * 1000